+++ pkg/metrics/state.go
@@ -7,6 +7,7 @@ const stateSubsystem = "state"
 type stateMetrics struct {
 	healthCheck             prometheus.Gauge
 	policerConsistencyState prometheus.Gauge
+	policerOptimalPlacement prometheus.Gauge
 }
 
 func newStateMetrics() stateMetrics {
@@ -23,12 +24,19 @@ func newStateMetrics() stateMetrics {
 			Name:      "policer_consistency_state",
 			Help:      "Current Policer's consistency state. 0 for inconsistent/unknown state; 1 for synchronized",
 		}),
+		policerOptimalPlacement: prometheus.NewGauge(prometheus.GaugeOpts{
+			Namespace: storageNodeNameSpace,
+			Subsystem: stateSubsystem,
+			Name:      "policer_optimal_placement",
+			Help:      "Current Policer's optimal placement state. 0 for non-optimal/unknown placement; 1 for optimal placement",
+		}),
 	}
 }
 
 func (m stateMetrics) register() {
 	prometheus.MustRegister(m.healthCheck)
 	prometheus.MustRegister(m.policerConsistencyState)
+	prometheus.MustRegister(m.policerOptimalPlacement)
 }
 
 func (m stateMetrics) SetHealth(s int32) {
@@ -40,5 +48,13 @@ func (m stateMetrics) SetPolicerConsistencyState(consistent bool) {
 		m.policerConsistencyState.Set(1)
 	} else {
 		m.policerConsistencyState.Set(0)
+	}
+}
+
+func (m stateMetrics) SetPolicerOptimalPlacementState(optimal bool) {
+	if optimal {
+		m.policerOptimalPlacement.Set(1)
+	} else {
+		m.policerOptimalPlacement.Set(0)
 	}
 }